api/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-decouple
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
email-validator
PyJWT
python-multipart
aiofiles
aiohttp
httpx
alembic
psycopg2-binary